refactor: GlobalStyle 폰트 설정 추가
