+++ src/app/_component/feature.css.ts
@@ -34,7 +34,6 @@ export const title = style({
   width: '940px',
   fontSize: '46px',
   fontWeight: '600',
-  fontFamily: 'SUIT-Regular',
   color: `${styleToken.color.gray1}`,
   transition: 'transform 1.0s ease-out',
   transform: 'translateY(100%)',
@@ -124,7 +123,6 @@ export const featureTitleContainer = style({
 export const featureTitle = style({
   fontSize: '30px',
   fontWeight: '700',
-  fontFamily: 'SUIT-Regular',
   color: `${styleToken.color.alert_success}`,
   textAlign: 'center',
   margin: '0 0 0 20px',
@@ -161,7 +159,6 @@ export const circle = style({
 export const featureDescription = style({
   fontSize: '18px',
   fontWeight: '100',
-  fontFamily: 'SUIT-Regular',
   color: `${styleToken.color.gray2}`,
   margin: '30px 0 0 8px',
   lineHeight: '32px',
