+++ src/app/_component/navigationHeader.css.ts
@@ -29,7 +29,6 @@ export const logoContainer = style({
   gap: '10px',
   fontSize: '20px',
   fontWeight: 'bold',
-  fontFamily: 'SUIT-Regular',
   color: `${styleToken.color.gray1}`,
   cursor: 'pointer',
 });
